-Removed some warnings -Fixed pfDrawManager.
git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler_1.0@1585 08b38cba-cd3b-11de-854e-f91c5b6e4272

+++ Kernel/vjDebug.cpp
@@ -97,7 +97,7 @@ void vjDebug::addAllowedCategory(int cat)
 bool vjDebug::isCategoryAllowed(int cat)
 {
    // If no entry for cat, grow the vector
-   if(mAllowedCategories.size() < (cat+1))
+   if((int)mAllowedCategories.size() < (cat+1))
       growAllowedCategoryVector(cat+1);
 
    // If I specified to listen to all OR
